Fix Irodori-TTS-Lite precision (upstream only accepts fp32/bf16)
Upstream irodori_tts.inference_runtime.resolve_runtime_dtype rejects
precision='fp16'; only fp32 and bf16 are valid. Lite's configure(
force_fp16=True) already forces fp16 inside the FusedInt4Linear kernel
path, so we let RuntimeKey use fp32 like the existing Irodori-TTS app.

+++ src/apps/irodori_lite_app.py
@@ -96,9 +96,9 @@ def get_runtime():
                 checkpoint=checkpoint_path,
                 model_device=MODEL_DEVICE,
                 codec_repo=CODEC_REPO,
-                model_precision="fp16",
+                model_precision="fp32",
                 codec_device=CODEC_DEVICE,
-                codec_precision="fp16",
+                codec_precision="fp32",
                 compile_model=False,
                 compile_dynamic=False,
             )
